Journals class now implements a method for cleaning each journal entry and a method to loop through all journals in the stream, clean them, then return the results. We still need to fill in the details of how to actually clean journals, and how to have multiple methods of cleaning the journals.
These changes to Journals class are part of an attempt to make it "play
nice" with JournalsManager which allows for parallel cleaning of the journals
files.

In the scripts folder are more examples of how to run some of the programs.

Any changes to programs in the parse_journal folder were because I changed
syntax or functionlity (slightyly) and needed to make sure all the tests
still passed.