Updated the journal cleaning files so that they use flags requesting to clean the journals for survival analysis.
I modified the journal class to also hold a variable called "features"
which can be a list of features describing a single journal, for example
[word count, number of sentences]

The survival analysis method for cleaning journals will print journal
features (i.e. the list in journal.features) to a file rather than
the text from the journal. Topic modeling still requires that the cleaned
journal text be printed to a file.